requirements.txt
Flask
Jinja2
numpy
pandas
matplotlib
praat-parselmouth
librosa
soundfile
scipy
reportlab
